+++ test/bfp_tests/src/bfp/real/test_bfp_use_exponent.c
@@ -125,7 +125,7 @@ TEST(bfp_use_exponent, bfp_s32_use_exponent)
       "end_exp: %d\n"
       "Expected: %ld * 2^(%d)  <--  (%ld >> %d) * 2^(%d + %d)  // was: %ld\n", 
       init_exp, end_exp,
-      expected[i], end_exp, dataA_copy[i], delta_p, init_exp, delta_p, dataA[i]);
+      (long int) expected[i], end_exp, (long int) dataA_copy[i], delta_p, init_exp, delta_p, (long int) dataA[i]);
       
   }
 }
