+++ test/vect_tests/src/vect/test_vect_add.c
@@ -454,7 +454,7 @@ TEST(vect_add, vect_s32_add_random)
 
         for(unsigned int i = 0; i < len; i++){
             int32_t expected = add_s32(B[i], C[i], b_shr, c_shr);
-            if(expected != A[i]) sprintf(msg_buff, sprintpat,v, i, len, A[i], B[i], b_shr, C[i], c_shr, (unsigned)A[i], (unsigned)B[i],  (unsigned)C[i]);
+            if(expected != A[i]) sprintf(msg_buff, sprintpat,v, i, len, (long unsigned int) A[i], (long unsigned int) B[i], b_shr, (long unsigned int) C[i], c_shr, (unsigned int) A[i],  (unsigned int) B[i],  (unsigned int) C[i]);
             TEST_ASSERT_EQUAL_MESSAGE(expected, A[i], msg_buff);
         }
         TEST_ASSERT_EQUAL(vect_s32_headroom(A, len), hr);
@@ -464,7 +464,7 @@ TEST(vect_add, vect_s32_add_random)
 
         for(unsigned int i = 0; i < len; i++){
             int32_t expected = add_s32(B[i], C[i], b_shr, c_shr);
-            if(expected != A[i]) sprintf(msg_buff, sprintpat,v, i, len, A[i], B[i], b_shr, C[i], c_shr, (unsigned)A[i],  (unsigned)B[i],  (unsigned)C[i]);
+            if(expected != A[i]) sprintf(msg_buff, sprintpat,v, i, len, (long unsigned int) A[i], (long unsigned int) B[i], b_shr, (long unsigned int) C[i], c_shr, (unsigned int) A[i],  (unsigned int) B[i],  (unsigned int) C[i]);
             TEST_ASSERT_EQUAL_MESSAGE(expected, A[i], msg_buff);
         }
         TEST_ASSERT_EQUAL(vect_s32_headroom(A, len), hr);
@@ -474,7 +474,7 @@ TEST(vect_add, vect_s32_add_random)
 
         for(unsigned int i = 0; i < len; i++){
             int32_t expected = add_s32(B[i], C[i], b_shr, c_shr);
-            if(expected != A[i]) sprintf(msg_buff, sprintpat,v, i, len, A[i], B[i], b_shr, C[i], c_shr, (unsigned)A[i],  (unsigned)B[i],  (unsigned)C[i]);
+            if(expected != A[i]) sprintf(msg_buff, sprintpat, v, i, len, (long unsigned int) A[i], (unsigned long int) B[i], b_shr, (unsigned long int) C[i], c_shr, (unsigned int) A[i],  (unsigned int) B[i],  (unsigned int) C[i]);
             TEST_ASSERT_EQUAL_MESSAGE(expected, A[i], msg_buff);
         }
         TEST_ASSERT_EQUAL(vect_s32_headroom(A, len), hr);
